+++ noodles-bam/CHANGELOG.md
@@ -18,6 +18,10 @@
 
 ### Removed
 
+  * bam/async/writer: Remove `Writer::write_sam_record`.
+
+    Use `Writer::write_alignment_record` instead.
+
   * bam/record/convert: Remove `Record::try_from_sam_record`.
 
   * bam/writer: Remove `Writer::write_sam_record`.